% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{active_databases}
\alias{active_databases}
\title{Retrieve a listing of active databases}
\usage{
active_databases(
  path = copernicus::copernicus_path(),
  include = c("none", "path", "database")[1]
)
}
\arguments{
\item{path}{chr, the root data directory}

\item{include}{chr, one or more variables to include, including "path" or
"database".  "none" (default) means to skip.}
}
\value{
data frame of active databases (region, group, period, product_id and
longname, with path and database possibly added)
}
\description{
This list, located in lut/active-databases.csv, is manually curated.
}
