% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andreas-package.R
\docType{package}
\name{andreas-package}
\alias{andreas-package}
\alias{andreas}
\title{andreas}
\description{
Serves Tandy Center for Ocean Forecasting Copernicus needs.
}
\details{
Andreas Copernicus was Nicolaus Copernicus' lesser known brother.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/andreas}
}

}
\author{
\strong{Maintainer}: Tupper Ben \email{btupper@bigelow.org}

Other contributors:
\itemize{
  \item Johnson Omi \email{ojohnson@bigelow.org} [contributor]
}

}
\keyword{internal}
