% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{archive_andreas}
\alias{archive_andreas}
\title{Archive a stars object to a database}
\usage{
archive_andreas(x, lut, path = ".", time = NULL, ...)
}
\arguments{
\item{x}{stars object}

\item{lut}{a tabular database (ala from \code{read_product_lut()})}

\item{path}{char, the data path}

\item{time}{NULL}

\item{...}{arguments for \code{\link[copernicus:generate_filename]{copernicus::generate_filename()}}}
}
\value{
tabular database as a tibble
}
\description{
Archive a stars object to a database
}
