% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{coerce_dates}
\alias{coerce_dates}
\title{Coerce user dates to fit within the catalog dates}
\usage{
coerce_dates(x = Sys.Date() + c(0, 9), y = read_product_lut())
}
\arguments{
\item{x}{vector of Dates}

\item{y}{a catalog table of one or more datasets}
}
\value{
two element Date class object corced to fit within the catalog data range
}
\description{
Coerce user dates to fit within the catalog dates
}
