% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{common_name}
\alias{common_name}
\title{Retrieve the common name of a given variable name}
\usage{
common_name(
  x = c("bottomT", "pbo", "sob", "tob", "mlotst", "so", "thetao", "uo", "vo", "wo",
    "zos"),
  lut = variable_lut()
)
}
\arguments{
\item{x}{chr, vector of one or more variable names}

\item{lut}{table of look up values}
}
\value{
charcater vector of short common names
}
\description{
Retrieve the common name of a given variable name
}
