% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_id.R
\name{decompose_dataset_id}
\alias{decompose_dataset_id}
\title{Decompose dataset_id into constituent parts}
\usage{
decompose_dataset_id(
  x = c("cmems_mod_glo_phy_anfc_0.083deg_P1D-m", "cmems_mod_glo_phy_my_0.083deg_P1D-m",
    "cmems_mod_glo_phy_myint_0.083deg_P1D-m"),
  col_names = c("origin", "group", "area", "theme", "type", "compinfo", "tempres")
)
}
\arguments{
\item{x}{chr one or more dataset ids}

\item{col_names}{chronicled, the names to use for the output table variables}
}
\value{
table of constituent parts
\itemize{
\item{origin}
\item{group}
\item{area}
\item{theme}
\item{type}
\item{compinfo}
\item{tempres}
\item{pref_order, the prefered ordering determined by 'type'}
}
}
\description{
Decompose dataset_id into constituent parts
}
\seealso{
\href{https://help.marine.copernicus.eu/en/articles/6820094-how-is-the-nomenclature-of-copernicus-marine-data-defined}{Copernicus Docs}
}
