% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{drop_degenerate}
\alias{drop_degenerate}
\title{Drop a degenerate dimension (usually depth or time)}
\usage{
drop_degenerate(x)
}
\arguments{
\item{x}{stars object}
}
\value{
stars object possibly with fewer dimensions
}
\description{
This has only narrow utility when downloading form Copernicus having filtered
to a single depth (which is retained).  This function will drop that.
}
