% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{extract_points}
\alias{extract_points}
\title{Extract point data from a stars object}
\usage{
extract_points(x, y, form = c("long", "wide"))
}
\arguments{
\item{x}{stars object (with or without time dimension)}

\item{y}{sf point data}

\item{form}{chr one of either "long" (the default) or "wide" to control 
output column layout.}

\item{...}{other arguments passed to `stars::st_extract()`}
}
\value{
table of variable data for each input point
}
\description{
Extract point data from a stars object
}
