% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_andreas}
\alias{fetch_andreas}
\title{A wrapper around \code{copernicus::fetch_copernicus()} to fetch Copernicus data as
stars objects.}
\usage{
fetch_andreas(
  x,
  drop_depth = TRUE,
  time = c(Sys.Date(), as.Date(max(x$end_time))),
  bb = c(xmin = -180, xmax = 180, ymin = -90, ymax = 90),
  coerce_time = TRUE,
  must_have_time = TRUE,
  verbose = FALSE,
  form = c("stars", "filename", "list")[1],
  ...
)
}
\arguments{
\item{x}{table - a look up table with one or more datasets}

\item{drop_depth}{logical, if TRUE then drop the depth dimension}

\item{time}{NULL or a two element range of dates}

\item{bb}{NULL or a 4-element vector for subsetting}

\item{coerce_time}{logical, if TRUE then coerce min and max time to fit within
the catalog offering}

\item{must_have_time}{logical, if TRUE and the raster does not have a time
dimension (ala only one day) then add one}

\item{verbose}{logical, if TRUE be verbose}

\item{form}{str, one of "stars" or "filename" to return a stars object or the filename
If "filename" then no attempt is made to read the downloaded file and \code{cleanup} is ignored.
If "list" then read the file into a list of stars objects where variables are grouped
by dimensional dependence.}

\item{...}{other arguments passed to \code{copernicus::fetch_copernicus()}}
}
\value{
a list of stars objects groups by dataset_id and depth.  Each element,
if not NULL, will have an "andreas" attribute that provides a list with
dataset_id, depth and time.  We use this because a degenerate dimension (just one
element) may be dropped, and so the time/depth context may be lost
}
\description{
A wrapper around \code{copernicus::fetch_copernicus()} to fetch Copernicus data as
stars objects.
}
