% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{fetch_static}
\alias{fetch_static}
\title{Fetch static data for a product suite}
\usage{
fetch_static(product_id = "GLOBAL_ANALYSISFORECAST_PHY_001_024")
}
\arguments{
\item{product_id}{chr the product id}
}
\value{
a stars object
}
\description{
Fetch static data for a product suite
}
