% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{format_depth_level}
\alias{format_depth_level}
\title{Farmat a depth as a depth name}
\usage{
format_depth_level(x)
}
\arguments{
\item{x}{num the actual depth(s)}
}
\value{
character - the deoth(s) formatted as a standard string
}
\description{
Farmat a depth as a depth name
}
