% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_extension}
\alias{get_extension}
\title{Retrieve the alphanumeric file extension}
\usage{
get_extension(x = c("a.tif", "b.nc", "c.txt", "d.j-p"), sep = ".")
}
\arguments{
\item{x}{chr one or more filenames}

\item{sep}{chr or NA, if character then prepend this to any extensions found}
}
\value{
character string
}
\description{
Retrieve the alphanumeric file extension
}
\seealso{
https://stackoverflow.com/questions/7779037/extract-file-extension-from-file-path
}
