% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize_databases}
\alias{harmonize_databases}
\title{Harmonize two databases (multiyear MY and analysis-forecast ANFC)}
\usage{
harmonize_databases(
  my_path = copernicus_path("jordanbasin/GLOBAL_MULTIYEAR_PHY_001_030"),
  anfc_path = copernicus_path("jordanbasin/GLOBAL_ANALYSISFORECAST_PHY_001_024"),
  dry_run = TRUE,
  verbose = !dry_run
)
}
\arguments{
\item{my_path}{chr the path to the multiyear database}

\item{anfc_path}{chr, the path to the anfc database}

\item{dry_run}{logical, if TRUE just compute the overlap by don't actually
purge data}

\item{verbose}{logical, if TRUE output messages about what is to be purged}
}
\value{
database of anfc data that has been purged (unless dry_run is TRUE)
}
\description{
If data are accruing for MY and ANFC at some point there may be overlap.  Newer
MY files have the same content (or intent) as the older ANFC.  This function will
purge ANFC files where they overlap with MY.  Our assumption is the MY files
have more vetted and stable data than ANFC.
}
