% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_databases}
\alias{list_databases}
\title{Retrieve a list of databases}
\usage{
list_databases(
  path = copernicus::copernicus_path(),
  pattern = "^database$",
  form = c("path", "table")[1]
)
}
\arguments{
\item{path}{chr, the root data directory}

\item{pattern}{chr, database filename regex pattern to search for}

\item{form}{one of "path" or "table"}
}
\value{
database paths relative to the root path
}
\description{
Retrieve a list of databases
}
