% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_files}
\alias{list_files}
\title{List files for a database}
\usage{
list_files(path, pattern = "*.tif", exclude = "static")
}
\arguments{
\item{path}{character the directory to catalog}

\item{pattern}{character, the filename pattern (as glob) to search for}

\item{exclude}{chr, one or more character patterns to exclude}
}
\value{
file list
}
\description{
List files for a database
}
