% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{list_luts}
\alias{list_luts}
\title{List the available LUTS}
\usage{
list_luts(path = copernicus::copernicus_path("lut"))
}
\arguments{
\item{path}{chr, the path to the LUT directory}
}
\value{
character vector of one of more LUT names
}
\description{
List the available LUTS
}
