% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{list_static}
\alias{list_static}
\title{List static variables in a regional product suite}
\usage{
list_static(
  path = copernicus::copernicus_path("gom", "GLOBAL_ANALYSISFORECAST_PHY_001_024")
)
}
\arguments{
\item{path}{chr, the copernicus data path}
}
\value{
the names of available static variables
}
\description{
List static variables in a regional product suite
}
