% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-databases.R
\name{merge_database}
\alias{merge_database}
\title{Merge compatible databases for one region}
\usage{
merge_database(path = copernicus_path("chfc"), rm_dups = TRUE, ...)
}
\arguments{
\item{path}{chr the path to the regional data repository}

\item{rm_dups}{logical, if TRUE remove duplicates preferentially 
retaining MULTIYEAR over ANALAYSISFORECAST products and chosing
"my" over "myint" datasets on a given date}

\item{...}{other arguments for [read_database]}
}
\value{
a merged database
}
\description{
Merge compatible databases for one region
}
