% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{read_buoys}
\alias{read_buoys}
\title{Read Gulf of Maine buoy data as CSV or sf}
\usage{
read_buoys(
  filename = system.file("exdata/buoy_listing.com", package = "andreas"),
  form = c("table", "sf")[2]
)
}
\arguments{
\item{filename}{chr the name of the file to read}

\item{form}{chr one of "table" or "sf" (default)}
}
\value{
eirther a data frame or spatial data frame
}
\description{
Read Gulf of Maine buoy data as CSV or sf
}
