% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_database}
\alias{read_database}
\title{Read a file-list database}
\usage{
read_database(path, multiple = FALSE, filename = "database", ...)
}
\arguments{
\item{path}{character the directory with the database}

\item{multiple}{logical, if TRUE then switch to a \link{merge_database} where
the path points to a directory that contains multiple compatible
databases.}

\item{filename}{character, optional filename}

\item{...}{optional arguments for \link{merge_database}}
}
\value{
a tibble (possibly empty if the database doesn't exist)
}
\description{
Read a file-list database
}
