% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{read_depth_lut}
\alias{read_depth_lut}
\title{Read a table of (standard) depth values}
\usage{
read_depth_lut(levels = 50, path = copernicus::copernicus_path("lut"))
}
\arguments{
\item{levels}{num the number of depth levels}

\item{path}{chr, the path to the LUT directory}
}
\value{
data frame with depth and name columns
}
\description{
Read a table of (standard) depth values
}
