% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{read_global_static}
\alias{read_global_static}
\title{Read global static variables, possibly subsetting with a bounding box}
\usage{
read_global_static(
  name = c("deptho", "mask"),
  product = "GLOBAL_ANALYSISFORECAST_PHY_001_024",
  bb = NULL
)
}
\arguments{
\item{name}{chr, the (names) of the static variables to read}

\item{product}{chr, the copernicus product_id}

\item{bb}{bounding box as a numeric vector, st_bbox, or an object from which a
bounding box can be derived. NULL to skip.}
}
\value{
stars object
}
\description{
Read global static variables, possibly subsetting with a bounding box
}
