% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-copernicus.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{copernicus_path}
\alias{make_path}
\alias{set_point}
\alias{bind_attrs}
\alias{bind_stars}
\alias{dataset_period}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{copernicus}{\code{\link[copernicus]{bind_attrs}}, \code{\link[copernicus]{bind_stars}}, \code{\link[copernicus]{copernicus_path}}, \code{\link[copernicus]{dataset_period}}, \code{\link[copernicus]{make_path}}, \code{\link[copernicus]{set_point}}}
}}

