% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{subset_static_by_region}
\alias{subset_static_by_region}
\title{A convenience function that subsets (or copies) the primary static arrays
(in `copernicus_path("static")`) to region specified.}
\usage{
subset_static_by_region(
  name = c("deptho", "mask", "deptho_lev"),
  regions = c("world", "nwa", "gom", "chfc"),
  bb = NULL
)
}
\arguments{
\item{name}{chr, one or more static variable names}

\item{regions}{chr one or more destination region name}

\item{bb}{bounding box - something from which an st_bbox can be made.
Ignored if region is "world", in which case we essentially copy to world. If 
not provided and the `cofbb` package is available, then retrieve that
dynamically}
}
\value{
NULL invisibly
}
\description{
A convenience function that subsets (or copies) the primary static arrays
(in `copernicus_path("static")`) to region specified.
}
