% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{variable_lut}
\alias{variable_lut}
\title{Retrieve a look up table (lut) relating variable names}
\usage{
variable_lut()
}
\arguments{
\item{table}{with "variable", "name", "longname" and "unit"}
}
\description{
Retrieve a look up table (lut) relating variable names
}
