% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{coerce_dates}
\alias{coerce_dates}
\title{Coerce user dates to fit within the catalog dates}
\usage{
coerce_dates(
  mydates = Sys.Date() + c(0, 9),
  catalog = read_product_lut(),
  warn = interactive()
)
}
\arguments{
\item{mydates}{vector of Dates}

\item{catalog}{a catalog table of one or more datasets}

\item{warn}{logcal, if TRUE thenissue a warning when the input is coerced}
}
\value{
two element Date class object corced to fit within the catalog data range
}
\description{
Coerce user dates to fit within the catalog dates
}
