% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{create_lut}
\alias{create_lut}
\title{Generate a LUT suitable for the package.}
\usage{
create_lut(
  product = "GLOBAL_ANALYSISFORECAST_BGC_001_028",
  region = "nowhere",
  catalog = nicolaus::read_catalog(),
  save_lut = FALSE
)
}
\arguments{
\item{product}{chr the name of the product}

\item{region}{chr, the name of the region}

\item{catalog}{table of products (unflattened)}

\item{save_lut}{log, if TRUE save to CSV format in \code{inst/lut}}
}
\value{
a table of look up values.  You'll edit this file to decide which to fetch
and what depths to fetch from.  Here's what is in the output table
\itemize{
\item product_id
\item title
\item dataset_id
\item dataset_name
\item short_name
\item standard_name
\item units
\item depth "sur"
\item fetch  "no"
\item mindepth 0
\item maxdepth 1
}
}
\description{
If you chose to save (see \code{save_lut}) then be advised existing an existing
lut will be overwritten.
}
