% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{create_lut}
\alias{create_lut}
\title{Generate a LUT suitable for the package.}
\usage{
create_lut(
  x = "GLOBAL_ANALYSISFORECAST_BGC_001_028",
  catalog = read_product_catalog(),
  save_lut = FALSE
)
}
\arguments{
\item{x}{chr the name of the product}

\item{catalog}{table of products (unflattened)}

\item{save_lut}{log, if TRUE save to CSV format in `inst/lut`}
}
\value{
a table of look up values.  You'll edit this file to decide which to fetch
 and what depths to fetch from.  Here's what is in the output table
 
* product_id
* title
* dataset_id
* dataset_name
* short_name
* standard_name
* units
* depth "sur"
* fetch  "no"
* mindepth 0
* maxdepth 1
}
\description{
If you chose to save (see `save_lut`) then be advised existing an existing
lut will be overwritten.
}
