% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_andreas}
\alias{fetch_andreas}
\title{A wrapper around [copernicus::fetch_copernicus()] to fetch Copernicus data as
stars objects.}
\usage{
fetch_andreas(x, drop_depth = TRUE, ...)
}
\arguments{
\item{x}{table - a look up table with one or more datasets}

\item{drop_depth}{logical, if TRUE then drop the depth dimension}

\item{...}{other arguments passed to [copernicus::fetch_copernicus()]}
}
\value{
stars object
}
\description{
A wrapper around [copernicus::fetch_copernicus()] to fetch Copernicus data as
stars objects.
}
