% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-databases.R
\name{merge_database}
\alias{merge_database}
\title{Merge compatible databases for one region}
\usage{
merge_database(path = copernicus_path("chfc"), dups = "latest", ...)
}
\arguments{
\item{path}{chr the path to the regional data repository}

\item{dups}{chr, control how to handle duplicate variables on a given day,
one of "latest" (most recent), "earliest", "all"}

\item{...}{other arguments for [read_database]}
}
\value{
a merged database
}
\description{
Merge compatible databases for one region
}
