% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_copernicus}
\alias{read_copernicus}
\alias{read_andreas}
\title{Read one or more copernicus files}
\usage{
read_copernicus(
  db,
  path,
  crs = 4326,
  bb = NULL,
  ext = ".tif",
  name_by = c("variable", "name", ".name")[3],
  group = name_by,
  tolerance = 1e-06,
  datetime = c("date", "posixt")[1],
  ...
)

read_andreas(db, path, ...)
}
\arguments{
\item{db}{tibble, database of selected records}

\item{path}{char, the path to the data set}

\item{crs}{value to set the CRS}

\item{bb}{NULL or and object used for cropping}

\item{ext}{chr the file extension, tyocially '.tif' but set to '.nc' for
ncdf fils that might have depth}

\item{name_by}{chr, one of "variable" or "name"  Individual variables
are named by either the database "variable" or "name" attribute.}

\item{tolerance}{num see \code{\link[stars:c.stars]{stars::c.stars()}}}

\item{datetime}{chr, one of "date" (default) or "posix" to control class of
the date/time dimension}

\item{...}{other arguments for \code{stars::st_crop}}
}
\value{
stars object or NULL
}
\description{
By default the function tries to return an object with (variables of x, y, time)
dimensions.  If multiple times are provided, then each var must be equally
represented.
}
