% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_copernicus}
\alias{read_copernicus}
\alias{read_andreas}
\title{Read one or more copernicus files}
\usage{
read_copernicus(db, path, crs = 4326, bb = NULL, ext = ".tif", ...)

read_andreas(db, path, ...)
}
\arguments{
\item{db}{tibble, database of selected records}

\item{path}{char, the path to the data set}

\item{crs}{value to set the CRS}

\item{bb}{NULL or and object used for cropping}

\item{ext}{chr the file extension, tyocially '.tif' but set to '.nc' for
ncdf fils that might have depth}

\item{...}{other arguments for \code{stars::st_crop}}
}
\value{
stars object
}
\description{
By default the function tries to return an object with (variables of x, y, time)
dimensions.  If multiple times are provided, then each var must be equally
represented.
}
