% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{read_product_lut}
\alias{read_product_lut}
\title{Read a product LUT used for searches and merge with a product description}
\usage{
read_product_lut(
  region = "chfc",
  product_id = "GLOBAL_ANALYSISFORECAST_PHY_001_024",
  description = nicolaus::read_catalog()
)
}
\arguments{
\item{region}{chr, the regional name}

\item{product_id}{chr, the product identifier.  The associated file must exist}

\item{description}{table, the table produced by \link[nicolaus:read_catalog]{nicolaus::read_catalog} for
the specified \code{product_id}}
}
\value{
a table with the following
\itemize{
\item product_id
\item dataset_id
\item dataset_name
\item name chr short_name converted to camelCase (used for file View(xstorage)
\item short_name
\item standard_name
\item units
\item depth really a depth name "sur", "bot", "mld", "zos" or NA but could be a number (as character())
\item fetch   "yes", "no" or NA
\item mindepth for \code{copernicusmarine} CLI request
\item maxdepth for \code{copernicusmarine} CLI request
\item start_time known data service availability
\item end_time known data service availability
}
}
\description{
Read a product LUT used for searches and merge with a product description
}
