% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.R
\name{read_static}
\alias{read_static}
\title{Read static variables for a regional dataset, possibly subsetting with a bounding box}
\usage{
read_static(
  name = c("deptho", "mask"),
  path = copernicus::copernicus_path("gom", "GLOBAL_ANALYSISFORECAST_PHY_001_024"),
  bb = NULL
)
}
\arguments{
\item{name}{chr, the (names) of the static variables to read}

\item{path}{chr, the copernicus data path}

\item{bb}{bounding box as a numeric vector, st_bbox, or an object from which a
bounding box can be derived. NULL to skip.}
}
\value{
stars object
}
\description{
\emph{N.B.} When combining these with other Copernicus layers using \code{c()} you may
encounter a message that says R doesn't know how to combine them.  Just increase
the \code{tolerance} value to \code{1e-6} and all will be well.  I can't explain why this
mismatch occurs, but this is the solution.
}
