% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_match}
\alias{st_match}
\title{Crop and warp a raster to match a template}
\usage{
st_match(x, y, ...)
}
\arguments{
\item{x}{stars object to be cropped and warped}

\item{y}{stars object serves as template}

\item{...}{other arguments for \link[stars:st_crop]{stars::st_crop} and \link[stars:st_warp]{stars::st_warp}}
}
\value{
stars object
}
\description{
Crop and warp a raster to match a template
}
