% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{tabulate_missing}
\alias{tabulate_missing}
\title{Tabulate missing data by '.name' and 'period'}
\usage{
tabulate_missing(x = read_database(copernicus_path("chfc"), multiple = TRUE))
}
\arguments{
\item{x}{a database}
}
\value{
data frame of metadata (id, .name, period), startdate (Date), enddate(Date) and missingdates (list)
}
\description{
Tabulate missing data by '.name' and 'period'
}
